requirements.txt
torch
torchvision
Pillow
pandas
opencv-python
tqdm